refactor(InitializeService): add service to init data
