+++ src/main/java/subway/controller/SubwayController.java
@@ -4,8 +4,8 @@ import java.util.Scanner;
 import java.util.function.Supplier;
 import subway.domain.Line;
 import subway.domain.LineRepository;
-import subway.domain.Station;
 import subway.domain.StationRepository;
+import subway.service.InitializeService;
 import subway.view.InputView;
 import subway.view.OutputView;
 
@@ -13,14 +13,16 @@ public class SubwayController {
 
     private final InputView inputView;
     private final OutputView outputView;
+    private final InitializeService initializeService;
 
     public SubwayController(Scanner scanner) {
         this.inputView = new InputView(scanner);
         this.outputView = new OutputView();
+        this.initializeService = new InitializeService();
     }
 
     public void run() {
-        initialSettings();
+        initializeService.initializeData();
 
         String firstChoiceStr;
         int firstChoice;
@@ -100,13 +102,8 @@ public class SubwayController {
         }
     }
 
-
-
-
-
-
     private void processToAddStation() {
-        createStation(inputView.inputStationNameToAdd());
+        initializeService.createStation(inputView.inputStationNameToAdd());
         outputView.successAddStation();
     }
 
@@ -133,7 +130,7 @@ public class SubwayController {
         String lineName = inputView.inputLineNameToAdd();
         String upbound = inputView.inputUpboundTerminusStation();
         String downbound = inputView.inputDownboundTerminusStation();
-        createLine(lineName, upbound, downbound);
+        initializeService.createLine(lineName, upbound, downbound);
         outputView.successAddLine();
     }
 
@@ -170,47 +167,6 @@ public class SubwayController {
         outputView.successDeleteSection();
     }
 
-    private void initialSettings() {
-        setStations();
-        setLines();
-        setSections();
-    }
-
-    private void setStations() {
-        createStation("교대역");
-        createStation("강남역");
-        createStation("역삼역");
-        createStation("남부터미널역");
-        createStation("양재역");
-        createStation("양재시민의숲역");
-        createStation("매봉역");
-    }
-
-    private void createStation(String stationName) {
-        StationRepository.addStation(new Station(stationName));
-    }
-
-    private void setLines() {
-        createLine("2호선", "교대역", "역삼역");
-        createLine("3호선", "교대역", "매봉역");
-        createLine("신분당선", "강남역", "양재시민의숲역");
-    }
-
-    private void createLine(String lineName, String upbound, String downbound) {
-        LineRepository.addLine(
-                new Line(lineName, StationRepository.findStation(upbound), StationRepository.findStation(downbound)));
-    }
-
-    private void setSections() {
-        addSections("2호선", "강남역");
-        addSections("3호선", "남부터미널역");
-        addSections("3호선", "양재역");
-        addSections("신분당선", "양재역");
-    }
-
-    private void addSections(String lineName, String StationName) {
-        LineRepository.findLine(lineName).addSection(StationRepository.findStation(StationName));
-    }
 
     private <T> T repeat(Supplier<T> inputReader) {
         try {
